backend/get_demographic_data/requirements.txt
fastapi>=0.112.0,<0.113.0
uvicorn[standard]>=0.29.0,<0.30.0
requests>=2.31.0
pandas>=2.2.2
python-dotenv>=1.0.1
google-generativeai>=0.3.2
